+++ CHANGELOG.md
@@ -17,7 +17,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
-*
+* Issue with background color and change of palette colors
+
+## [0.5.5] - 2022-11-17
+
+### Fixed
+
+* PPU issue related to the maximum number of objects/sprite per line being 10, issue detected by ACID test
+* Object pixel drawing priority issue, issue detected by ACID test
+* Issue associated with the wrongful flipping of 8x16 sprites, issue detected by ACID test
+* Issue associated with drawing of window tiles, due to extra `update_stat()` operations, issue detected by ACID test
 
 ## [0.5.4] - 2022-11-15